08_function_calling/mcp_fastapi/requirements.txt
fastapi
uvicorn
pandas
requests
python-dotenv
